requirements.txt
Flask
Flask-Cors
google-cloud-aiplatform
python-dotenv
google-auth
vertexai
gunicorn